Move mutex functionality into BaseMutexFileAppender
Most of the subclasses of BaseFileAppender are using only a subset of
its functionality. There are only two classes require the mutex
functionality. Therefore, the mutex functionality is moved into a new
abstract class (BaseMutexFileAppender) to re-enforce the single
responsibility principle.
