+++ src/NLog/Internal/FileAppenders/MutexMultiProcessFileAppender.cs
@@ -58,7 +58,7 @@ namespace NLog.Internal.FileAppenders
     /// (global named mutex is used for this)
     /// </remarks>
     [SecuritySafeCritical]
-    internal class MutexMultiProcessFileAppender : BaseFileAppender
+    internal class MutexMultiProcessFileAppender : BaseMutexFileAppender
     {
         public static readonly IFileAppenderFactory TheFactory = new Factory();
