+++ src/NLog/Internal/FileAppenders/RetryingMultiProcessFileAppender.cs
@@ -48,7 +48,7 @@ namespace NLog.Internal.FileAppenders
     /// to get exclusive write access and retries if it's not available.
     /// </summary>
     [SecuritySafeCritical]
-    internal class RetryingMultiProcessFileAppender : BaseFileAppender
+    internal class RetryingMultiProcessFileAppender : BaseMutexFileAppender
     {
         public static readonly IFileAppenderFactory TheFactory = new Factory();
